requirements.txt
Flask>=2.3,<4
python-dotenv>=1.0
gunicorn>=21
Flask-SQLAlchemy
Flask-Bcrypt
Flask-Login
Flask-Migrate
Flask-WTF
requests
stripe
email-validator    # needed if your forms use EmailField / Email()
psycopg2-binary  # uncomment if you deploy with Postgres on Rendergunicorn

Flask-Admin
